requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.46
psycopg2-binary==2.9.7
pydantic==2.7.1
email-validator==2.3.0
python-dotenv==1.0.0
openai==1.28.2
requests>=2.32.2
python-multipart>=0.0.7
pdfplumber==0.11.0
pytesseract==0.3.13
pdf2image==1.17.0
pillow==10.4.0
jinja2==3.1.4
